`Command#writeTo`: Use the optional `callback` parameter of `Socket#write`.
`write` returns `true` if the data was successfully written to the socket, or
`false` if it was queued. In the latter case, the `callback` is invoked once
the queue has been flushed.

`Command#writeTo` now returns a promise that is fulfilled once the command has
been fully written to the socket. This avoids requiring the `Client` to keep
track of whether the connection is buffering.